Fix sonar warning
