+++ src/server/common/helpers/session-cache/site-utils.js
@@ -1,21 +1,6 @@
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
 import { getExemptionCache } from '~/src/server/common/helpers/session-cache/utils.js'
 import { getSiteNumber } from '~/src/server/exemption/site-details/utils/site-number.js'
-
-/**
- * Gets site details from exemption data with safe array access
- * @param { object } exemption - Exemption object
- * @param { number } siteIndex - Index of site to retrieve (defaults to 0 for single site flows)
- * @returns { object } Site details object or empty object if not found
- */
-export const getSiteDetailsBySite = (exemption, siteIndex = 0) => {
-  const { multipleSiteDetails } = exemption
-
-  if (!multipleSiteDetails?.multipleSitesEnabled) {
-    return exemption.siteDetails?.[0] ?? {}
-  }
-
-  return exemption.siteDetails?.[siteIndex] ?? {}
-}
 
 /**
  * Gets site data for the current page
