+++ src/server/common/helpers/session-cache/utils.js
@@ -1,6 +1,5 @@
 import { clone } from '@hapi/hoek'
-import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-utils.js'
-
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
 export const EXEMPTION_CACHE_KEY = 'exemption'
 
 /**
