+++ src/server/exemption/site-details/centre-coordinates/controller.js
@@ -2,8 +2,8 @@ import {
   getExemptionCache,
   updateExemptionSiteDetails
 } from '~/src/server/common/helpers/session-cache/utils.js'
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
 import {
-  getSiteDetailsBySite,
   setSiteData,
   setSiteDataPreHandler
 } from '~/src/server/common/helpers/session-cache/site-utils.js'
