+++ src/server/exemption/site-details/coordinate-system/controller.js
@@ -3,10 +3,10 @@ import {
   updateExemptionSiteDetails
 } from '~/src/server/common/helpers/session-cache/utils.js'
 import {
-  getSiteDetailsBySite,
   setSiteData,
   setSiteDataPreHandler
 } from '~/src/server/common/helpers/session-cache/site-utils.js'
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
 import {
   errorDescriptionByFieldName,
   mapErrorsForDisplay
