+++ src/server/exemption/site-details/coordinates-type/controller.js
@@ -2,10 +2,8 @@ import {
   getExemptionCache,
   updateExemptionSiteDetails
 } from '~/src/server/common/helpers/session-cache/utils.js'
-import {
-  getSiteDetailsBySite,
-  setSiteDataPreHandler
-} from '~/src/server/common/helpers/session-cache/site-utils.js'
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
+import { setSiteDataPreHandler } from '~/src/server/common/helpers/session-cache/site-utils.js'
 import {
   errorDescriptionByFieldName,
   mapErrorsForDisplay
