+++ src/server/exemption/site-details/enter-multiple-coordinates/controller.js
@@ -7,10 +7,8 @@ import {
   getExemptionCache,
   updateExemptionSiteDetails
 } from '~/src/server/common/helpers/session-cache/utils.js'
-import {
-  getSiteDetailsBySite,
-  setSiteDataPreHandler
-} from '~/src/server/common/helpers/session-cache/site-utils.js'
+import { setSiteDataPreHandler } from '~/src/server/common/helpers/session-cache/site-utils.js'
+import { getSiteDetailsBySite } from '~/src/server/common/helpers/session-cache/site-details-utils.js'
 import { getCoordinateSystem } from '~/src/server/common/helpers/coordinate-utils.js'
 import {
   MULTIPLE_COORDINATES_VIEW_ROUTES,
